JPA Repository + H2 db

+++ todo/restful-web-services/src/main/java/org/agcodes/restful_web_services/service/TodoHardcodedService.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Service;
 public class TodoHardcodedService {
 
   private static List<Todo> todos = new ArrayList<>();
-  private static int idCounter = 0;
+  private static Long idCounter = 0L;
 
   static {
     todos.add(new Todo(++idCounter, "test", "Learn AWS", new Date(), false));
